A more modular way to compile the HMAC functionality
